move logs language to a separate folder
