+++ public/app/plugins/datasource/cloudwatch/cloudwatch-logs/syntax.ts
